Fix corner cases test
The change is in the starting column of one row. We do not include the
top level function call in the the srcref. For example a function:

f <- function(x) x

will have one counter - x - with srcref starting at the start of x.

              filename functions first_line first_byte last_line
BEFORE: corner-cases.R       ANY         56         20        56
AFTER:  corner-cases.R       ANY         56         32        56
  last_byte first_column last_column first_parsed last_parsed value
         54           20          54           56          56     0
         54           32          54           56          56     0